+++ src/auth/clients/ClientInterface.php
@@ -5,11 +5,7 @@ namespace davidhirtz\yii2\skeleton\auth\clients;
 use davidhirtz\yii2\skeleton\models\AuthClient;
 
 /**
- * Interface ClientInterface.
- * @package davidhirtz\yii2\skeleton\auth\clients
- *
- * @property-read array $safeUserAttributes
- * @see ClientInterface::getSafeUserAttributes
+ * @property-read array $safeUserAttributes {@see ClientInterface::getSafeUserAttributes()}
  */
 interface ClientInterface extends \yii\authclient\ClientInterface
 {