+++ src/models/traits/SignupEmailTrait.php
@@ -4,15 +4,12 @@ namespace davidhirtz\yii2\skeleton\models\traits;
 
 use Yii;
 
-/**
- * @property string $email
- */
 trait SignupEmailTrait
 {
     public function sendSignupEmail(): void
     {
         $mail = Yii::$app->getMailer()->compose('@skeleton/mail/account/create', [
-            'user' => $this,
+            'user' => $this->user,
         ]);
 
         $mail->setSubject(Yii::t('skeleton', 'Sign up confirmation'))