+++ tests/unit/behaviors/TrailBehaviorTest.php
@@ -280,6 +280,9 @@ class TrailActiveRecord extends ActiveRecord
         ];
     }
 
+    /**
+     * @return UserQuery<User>
+     */
     public function getUser(): UserQuery
     {
         /** @var UserQuery $query */